avoid corrupting the version cached in redis
bulkGetOpsSince would corrupt the cached version key for a document if
called with a from version from the future. This is bad, because
confused clients can get redis into a permanently corrupt state.
Instead of relying on the from being a valid version, get the version
from the oplog db method each time