Fix flaky pollDebounce tests
The pollDebounce tests insert 10 documents and make sure
that debouncing leads to two separate chunks of documents
being passed through the 'insert' callback on the query.

The problem was that we weren't waiting for the document
creation to be submitted before sending the next document.

Even though the original test behavior was a more accurate
representation of what a real client would do, it lead to
unpredictable results. The timing of when the op was submitted
could have lead to all of the documents being added in one single
chunk, rather than two.