In memory db tests, test this package rather than a dependency
Tests now use sharedb-mingo-memory@0.1.2 which requires the user
to explicitly pass in a version of MemoryDB from which ShareDBMingo
extends. Prior to this change, sharedb-mingo-memory extended from
a version of MemoryDB from sharedb installed as a dependency
of sharedb-mingo-memory. That meant we were testing the latest
published version of sharedb, rather than the one in this
repository.

This change should resolve the decrease in test coverage here:
https://coveralls.io/builds/5983381