Add support for fetching a particular version of a snapshot
This change follows on from discussion in [this issue][1]. Its primary
aim is to allow clients to fetch an historical version of a document
(a snapshot), by providing either:

  - a version `number`, or
  - a desired `Date`

The entry-point for this feature is added to the `Connection` class,
deliberately separate from the `Doc`, because `Doc` is concerned with
"live" document actions such as subscribing and submitting ops,
whereas fetching an historical version of a document should not be
associated with these ideas.

The feature is called with:

```javascript
connection.getSnapshot(collection, id, version, callback): void;
```

The details of the interface are detailed in the README, and in the
code documentation.

This change includes support for projections, and use of the
`readSnapshots` middleware. It also hooks into `Connection`'s
`hasPending` method.

Performance optimisations are deemed out-of-scope for this change (see
the [issue][1] for more details).

Note that this change also adds a development dependency on [`lolex`][2]
which is used for mocking the time.

[1]: https://github.com/share/sharedb/issues/218
[2]: https://github.com/sinonjs/lolex