Increment fetched snapshot version
The version attached to an operation in the database is the version
before the operation has been applied to (or the version to which the
operation can be applied).

That means that when applying op v1, the document is v1 before the op
has been applied, and then v2 after the op has been applied.

This change updates the `getSnapshot` method to return snapshots with
versions consistent with this (ie the snapshots are now 1-based instead
of 0-based).