Promisify client-facing Doc methods
This change makes client-facing `Doc` methods return a `Promise` when
no `callback` is provided.

In order to do this, a `promisify` method is added to `util`, which will
return a `{ promise, callback }` object. Calling `callback()` will also
return the `promise` so that early `function` exits are possible with
`return callback();`