Add Snapshot class
This change adds a very simple `Snapshot` class, which currently has
no methods attached to it. Its primary use at the moment is to contain
the response to a snapshot request, and for use in `MemoryDB` in place
of `MemorySnapshot`.

As part of this change, the structure of the snapshot request return is
slightly modified to fit this class, including renaming of `version` to
`v`, and moving of `timestamp` to `m.ts`.