Remove snapshot fetch at time
This change makes a couple of review markups. The biggest is the
removal of the ability to fetch a snapshot at a given time. It has been
agreed that this would be a useful API, but we need to think about the
ramifications of exposing this API before doing so (eg adding more
indexes on database adapters for lookup by time, etc.).

In order to move the core functionality forward - that is the ability to
fetch a snapshot by version - this change removes the ability to fetch
by time, and it is left (for now) to the consumer to look up the version
they need themselves.

Alongside this change, `getSnapshot` has also been renamed to
`fetchSnapshot` to imply that we touch the database layer, and the
`Snapshot` class has had its `collections` property removed to be
consistent with database adapters, whilst also making `m` a property
that is always present.