Move milestone snapshots into their own database adapter
The concept of milestone snapshots can be completely decoupled from the
existing database adapter. This change reflects that concept in code by
extracting the milestone snapshot methods from the core database
adapter and into their own `MilestoneDB` class.

This means that a milestone database could be implemented using a
completely separate database from the primary ShareDB database.

As well as this change, some tweaks to the interface have been made:

  - Saving a snapshot does not require a callback. It emits an error if
    no callback is provided. This allows us to make saving a milestone
    on commit non-blocking, which avoid slowing down commits.
  - No default interval is provided for storing milestones. A default
    could be implemented by any adapters extending the base class, which
    can choose different defaults depending on the implementation.