Allow milestone snapshot saving logic to be overridden in the middleware
This change adds a flag to `SubmitRequest`: `saveMilestoneSnapshot`.
This flag defaults to `null`. If this value is left as `null`, then
milestone snapshots will be written according to the interval set in the
options passed into the provided `MilestoneDB` instance.

The value can be overridden in middleware, like in the test case in this
commit. If the value of `request.saveMilestoneSnapshot` is changed to
a `boolean`, then all interval logic is ignored, and the value of that
flag is used directly to determine if a milestone should be saved.

This allows more complex milestone logic, such as:

  - saving a milestone once per day
  - different intervals depending on collection
  - non-uniform milestone intervals