Expose `fetchSnapshot` requested version in `readSnapshots` middleware
In the `readSnapshots` middleware, it's currently impossible to
determine whether a snapshot is being returned as the result of a
`fetch` or a `fetchSnaphot`.

This change adds a version `v` field to the `request` when the
middleware is invoked from `fetchSnapshot`. This field is populated with
the value of the requested version, and can be used to determine that
the snapshot is the result of a `fetchSnapshot` rather than a `fetch`.