Add a `NoOpMilestoneDB` implementation
This change updates our `MilestoneDB` base class to be an empty shell
that throws whenever any of its methods are called in order to help
implementers.

We then also introduce a `NoOpMilestoneDB` implementation, which
implements all the methods, but does nothing. This is so that we can
drop this implementation into the `Backend` as a sensible default that
does nothing at all when called (instead of eg using `MemoryMilestoneDB`
and having the surprising behaviour of working until you restart your
application).

As part of this, the base call signature of the `saveMilestoneSnapshot`
callback is changed to remove the `wasSaved` flag. A snapshot should be
assumed to have saved if no error is returned in that callback.