Commonise valid version logic
This change introduces a `util.isValidVersion` method, which is common
between `SnapshotRequest` and `MilestoneDB`, and makes sure we're
consistent in what constitutes a valid version.

We check that versions are either `null` or a positive integer. We
reject `undefined`, based on the idea that `null` means that you meant
to set something to `null`, but with `undefined` that may not be the
case. It's also consistent with the version that can be passed into
`getOps` (ie `null` has the special meaning of "please give me the
latest version").