Rename Snapshot middleware "historical" flag to "byVersion"
Calling a snapshot "historical" might lead people to assume that it's
not the current version, which may not be the case (for example, if you
fetch the current version by its ID with `fetchSnapshot` or, pass in
a version of `null`).

This change updates the snapshot middleware to use the more descriptive
`byVersion`, which indicates that the snapshot was fetched by version.