Add test for re-subscribe memory leak
When a doc subscription stream ends, it's supposed to be removed from `agent.subscribedDocs[collection][docId]`.

When subscribing to a doc that's already subscribed to, the old stream gets closed, and the new stream gets set onto the above location, which acts as removing the old stream. However, prior to commit 41c40f6, the new stream actually got removed, which means the agent loses its reference to the stream.

This test for that scenario fails prior to 41c40f6 and passes with 41c40f6.