Debounce the fetch to prevent multiple requests
In some case, multiple fetch requests can be triggered. We want to prevent
to have to many of them wich will lead to unnecessary treatments and data flux.
If a rollback is trigger during the handleOp, all the messages received by the
open socket will trigger a fetch. We want to ignore them to acknowledge the
inflightFetch