Allow options to be passed for `fetch` and `getOps`
The database adapters allow an `options` object to be passed through to
them for enabling the return of metadata with a snapshot or ops.

Consumers could query the database directly, or even use the database
adapters, but this may give inconsistent results when comparing ops with
those fetched through `getOps`. For example, the Mongo adapter makes
sure that a valid set of ops with unique versions are returned, which
may not be the case when querying the database directly. Fetching ops
and snapshots through `Backend` methods also ensures that we call the
appropriate "sanitize" methods, and trigger the corresponding
middleware.

However, we don't expose this on `Backend.getOps` or `Backend.fetch`.
This change adds an optional `options` argument to these methods, which
can then be used to ask for metadata.

Note that an options argument has been added to `Backend.subscribe`, but
using it will return an error. This is to keep the signature consistent
with `fetch` and `getOps`. However, the implementation is beyond the
scope of this change, because we'd need to add some way to configure
`SubmitRequest.commit` to optionally pass metadata to the appropriate
clients, who provided that given option on `subscribe`.