Silence logger in tests
At the moment, the logger will log to console during tests, which is
noisy and undesirable.

This change adds a test setup file, which silences the logger by setting
all of its methods to no-ops.

This behaviour can be overridden with the `LOGGING` environment
variable:

```bash
LOGGING=true npm test
```