Use a presence version instead of `seq`
At the moment, we use `seq` as a proxy for "presence version". However,
this is problematic, because the `seq` is currently reset every time a
client reconnects, and we shouldn't really be checking against it.

This change does away with this usage and simply introduces a notion
`pv` presenve version field onto the presence message, which can be
used to reliably check the version of a presence with a given ID
(assuming a client's ID is truly unique to that session).