Add basic handshake negotiation to connection
This change updates how the client and server communicate when first
connecting, in a non-breaking way.

## Current behaviour

  - client creates and connects to a socket
  - client constructs a `Connection` object with the socket
  - the backend binds to the socket and constructs an `Agent`
  - the `Agent` sends an `init` message down the socket when constructed
  - the `Connection` receives the `init` message and initialises itself
    using information in the message

## Motivation for change

There are currently a few things that would be nice to "negotiate" when
a client connects. For example, clients may wish to:

  - the same `src` value when reconnecting, so that it persists as a
    stable "session ID"
  - set their default type
  - establish the exact versions of OT types being used
  - establish plugins and versions to use

More detail can be found here: https://github.com/share/sharedb/issues/337

## New behaviour

  - client creates and connects to a socket
  - client constructs a `Connection` object with the socket
  - the backend binds to the socket and constructs an `Agent`
  - the `Agent` sends an `init` message down the socket when constructed
  - **new behaviour starts here**
  - the `Connection` checks the `init` message to see if the backend is
    capable of handshaking
  - if the backend _cannot_ handshake (ie it's old), then the client
    initialises itself as before
  - if the backend _can_ handshake, then it will disregard the `init`
    message, and send its own `hs` "handshake" message to the server
  - the backend receives the handshake message and sends a response
  - the `Connection` receives the new `hs` message and initialises itself
    using information in the message (like before, but with more or
    different information information)

## Other changes

No changes have been made to the information being shared. In other
words, this commit only updates the way in which the backend and the
client communicate when first connecting.

As a convenience for some tests, which had to be tweaked (because they
assumed the connection was established immediately), this change also
adds an optional `callback` parameter to `Backend.connect`, which will
be invoked once the connection handshake completes.