Persist `Connection` id on reconnection
Right now, whenever a `Connection` is disconnected it:

  - sets its ID to `null`
  - resets its `seq` to `1`
  - clears its `Agent`

This makes it difficult for the server, and for remote clients, to keep
track of who is represented by a given ID (or `src`).

This change uses the new handshake protocol to allow the client to keep
its existing ID, so that it can persist even across a reconnection.