Add guards for large `seq`
Previously, `seq` would get reset every time a client reconnects.
Whilst not entirely robust, it did help to mitigate the possibility that
very long-lived connections might hit an integer overflow when
incrementing their `seq`.

This change adds a guard against this case, and throws a sensible error,
both on the client, and on the server.

Note that we cannot rely on the client's `Number.MAX_SAFE_INTEGER`,
because it is not supported by Internet Explorer.