Remove `util.callInSeries`
The hand-rolled `util.callInSeries` method was added without realising
that the `async` library was already in use in the library.

This change removes our custom method, and simply uses the appropriate
`async` functions instead.

We also add a utility `errorHandler` function that will call `done` in a
test if an error occurs, but not otherwise, which is useful when using
callback methods in a fire-and-forget fashion.