Add `Agent.src` field
This change adds an `Agent.src` field, which will accept a
client-provided ID. This is kept separate from the Agent-generated
`clientId` so that the agent's ID does not change during the handshake
(which could confuse any consumers relying on a stable client ID in the
`connect` middleware).

Any internal references to `clientId` then give precedence to the
consumer-provided `src`. If it hasn't been set (ie we're dealing with a
legacy client), then we fall back to using `clientId`.