Use `Agent._src` getter for presence `src`
This change updates the Presence `src` field to use the Agent's `_src`
getter, which will try to use a client-proposed ID, and then fall back
to the server-generated `clientId`.

This means that the `src` should be consistent across client
reconnections, which helps `RemotePresence` to determine which ops have
come from which presence providers.