Add a presence feature flag
This change adds a presence feature flag, so that consumers have to
actively opt in to use presence. This means that servers that don't want
to worry about potential fan-out issues don't need to worry about it.

Note that the feature flag will disable presence subscription and also
disable presence broadcasting, but notably _does not_ disable
unsubscribing from presence, so that clients may unsubscribe if the flag
is toggled while the server is running.