Fix flaky test
At the moment, we have a test that has a race condition. We submit ops
synchronously from two different connections, but it's not well defined
which of these will be committed first.

This change waits for the first submission to be acknowledged before
firing the second submit op. We still test the server `transformOp`,
because `doc2` is not subscribed (and hence submits an op at a version
behind the current snapshot version, which will need transforming).

We also fix the fact that the two ops previously did not combine into a
no-op by just firing off a redundant prop removal as the second op.