Ignore doc does not exist on presence destroy
When we destroy local presence, we attempt to update the presence to
`null` for remote clients as part of the cleanup.

However, if we delete a doc, and then destroy the associated presence,
the presence attempts to send a message on a deleted doc, which returns
an error.

Let's just ignore this error. There's no need to reset presence on a
deleted doc, because remote clients will [already assume it's `null`][1].

[1]: https://github.com/share/sharedb/blob/a81e2c99af77da12885f0fa860e87cc2602a6cc1/lib/client/presence/local-doc-presence.js#L90-L96