Amend submit logic
This change amends the `destroy` logic for `LocalDocPresence`, which
handled the error in the wrong place, and meant that the `LocalPresence`
would not correctly tidy up the reference to the destroyed local doc
presence.

This change adds a check to `LocalDocPresence.submit`. If we're
submitting `null` on a deleted doc, we just immediately return, since
the remote client will already assume a `null` presence in this case.