Catch errors thrown in `ot.applyOps`
The `ot.applyOps` utility function directly calls type methods, such as
`type.apply`. These type methods can potentially `throw`. This results
in an uncaught error, which we should avoid.

This change wraps the `ot.applyOps` functionality in a `try`/`catch`
block and returns the error, so that it can be handled appropriately.
We add a new `ERR_OT_OP_NOT_APPLIED` error code for this case.