Fix unit test that breaks in Node v14
Node v14 made some changes to `Stream` internals, which seems to break
this unit test with the new handshake approach.

This is because the client sends a handshake message just as the stream
is closing.

Since the stream is closed, the message can't be delivered, and -
according to the [docs][1]:

> The 'end' event **will not be emitted** unless the data is completely
> consumed.

Since this message isn't consumed, the `'end'` event is never emitted,
and this test fails.

Ideally we'd just listen for the `'close'` event, which should - in
theory - always be called, but Node versions before 14 have flaky
behaviour with this.

Instead, we copy the [behaviour][2] in `Agent` itself, and listen for
both `'end'` *and* `'close'` events.

[1]: https://nodejs.org/docs/latest-v12.x/api/stream.html#stream_event_end
[2]: https://github.com/share/sharedb/blob/061c421063388306a0c8f1c6e1dc9c450ab5a413/lib/agent.js#L351-L352