Fix flaky test
This change is a speculative fix for a flaky test. The `sharedb-mongo`
build has sometimes been failing because the following test sometimes
[fails][1]:

```
subscribe bulk with readSnapshots rejectSnapshotRead special ignorable error
```

The test fails in the `after each` hook, because we try to write a
snapshot to Mongo after Mongo has already been closed (ie `done()` has
already been called).

A potential cause for this is the fact that we don't wait for the
`submitOp` calls to complete before exiting the test. It's possible that
in a fraction of runs, the following happens:

 - Submit op on `fido`
 - Submit op on `spot`
 - `spot` op committed
 - `'receive'` called
 - nextTick, but `fido` op **still** not committed yet
 - Run assertions, and call `done()`
 - Mongo closed in the `afterEach` hook
 - `fido` finally tries to commit, but Mongo has been closed already

This change attempts to guard against this case by wrapping everything
in a `parallel` call, so we wait for the `submitOp` calls to complete
before exiting the test.

[1]: https://github.com/share/sharedb-mongo/runs/1565317742?check_suite_focus=true