Optionally send `source` to the server
This change adds the capability to send an op's `source` to the server,
for access in middleware.

# Motivation

The use-case that inspired this change was the ability to record some
usage stats based on users' ops. It's possible to record these stats
through separate HTTP calls, but recording stats through ShareDB would:

 - keep stats in sync with committed ops
 - allow servers to dictate stat definitions (enabling immediate
   roll-out; consistency; etc.)
 - leverage existing socket connections, which prevents having to open
   another socket; make lots of HTTP calls; or resort to
   [`sendBeacon`][1]

However, the type of stats we wish to record require more transient
information about the source of the op (eg keyboard; clipboard; mouse;
etc.).

It's clearly impractical to set this at connection time (where we set
other more "permanent" metadata, such as user ID), so we need some way
of signalling transient metadata attached to individual ops.

This can be communicated within the same client using the `source`
argument on `submitOp`, but this information never makes it to the
server.

# Implementation

## Usage

This new feature is opt-in, and is enabled on a per-`Doc` basis by
setting `doc.submitSource = true`.

No further configuration is required, and `source` can be set as it was
before in `submitOp`.

The `source` will be available in middleware at `request.extra.source`.

Of particular note is the fact that this feature supports complex
objects as `source`s, so it's possible to send complex metadata
structures to the server (although please note the side-effect mentioned
below).

## `op.x`

This change adds a new `x` property to the structure of a wrapped `op`.
This `x` represents "extra" information, which will not be committed to
the database, but will be made available in middleware (if consumers
want to commit it themselves, they are free to move data from `x` to `m`
in the middleware).

For now, the only property `x` can have is `source`, which will be
transmitted from the client.

## Side-effect

When using this feature, there is a necessary side-effect: any ops that
don't have a matching `source` will not be composed (since we don't know
how to disentangle these ops in the middleware).

This should be unimportant in the majority of cases (particularly if
`source` has not been actively set).

[1]: https://volument.com/blog/sendbeacon-is-broken